requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-ffmpeg-video-streaming==0.1.15
neo4j==5.18.0
python-dotenv==1.0.0 